refactor: share the model binding of Path and ScimFilter
Both are strings whose attribute names only mean something against a model, and
both carried their own subscription, class cache and pydantic plumbing. They
now derive from _BoundToModels, which keeps the generated class name and the
error messages each of them used to build by hand, and turns whatever either
gets wrong into a pydantic error, where only a path used to name the field it
was read from and carry its scimType.
